add refs to block
Add the idea of a reference count to the block node.  Also, first
check if a bock already exists for the given data.  If so, there is no
reason to recreate it, but instead it's ref count should be
incremented.  This might also have the nice side effect of improving
efficiency when writing files that have a lot in common.