implement bury function
The bury function transverses the list of dead nodes (i.e. those in
the dead bucket) and deletes them from the node bucket.  However, from
the time when the node was declared dead to the time of bury being
called the node may have been resurrected (i.e. a node with the same
original data).  The bury function will account for this and _not_
delete the node from the bucket--it will have to go back through the
reaping/burying process just like it did before.