more efficient reap
An attempt to make reap more efficient by not decrementing the ref
count.  A count of 1 should symbolize a dead node because we know it's
parent it being reaped as well.

Had to add a ref modification to bury to make sure that the count is
decremented in the case where the parent dies but the child doesn't.