ignore non-existent objs during gc
Since the mapred jobs that drive gc can timeout there is a possibility
that it won't delete all the nodes.  Since gc is performed from the
bottom-up that means that there will be nodes which potentially
reference non-existent nodes.  In order to avoid an exception during
the next gc invocation it must account for these dangling references
and safely ignore them.