fix race condition
If new data is being streamed in as GC is running there is a race
condition where a node will be falsely considered an orphan.  E.g., GC
adds a tally for each root pointed to from the TLDs and then moves
into the second phase of GC in the luwak_node namespace and at the
same time a new file is written.  In this case, the new root will be
considered orphaned if GC comes across it.