fix race condition...try 3
Not only do I not want to delete anything older than "now," but I also
don't want to delete anything created recently.  This is because one
(or many) process(es) could be writing a large file with many blocks
and those blocks will be written before the inner nodes are created or
the root is updated and GC will consider them as orphans.  Luwak won't
notice this and will update the root with the newely created and
promptly deleted node(s).  Then, when a process tries to retrieve the
data it will get a 'notfound'.

There is certainly a saner way to do this with the vector clock for
each node/block but I'm short on time and I'm not sure if it would be
as efficient.  I'm still experimenting here, after all.